Don't add invalid link relations to HAL output
HAL spec says that the href property of a link relation is
required. Certain parsers, thus, will break if a nil href is
included in the spec. This PR simply checks to make sure these
don't show up in the JSON output if the link you pass in happens
to be nil.