Support for multiple schema blocks
Either within the same class definition, or inheritable across the class
hierarchy. For example a child serializer which allows us to append/override previously defined schema attributes.

Depending on the adaptor in question, existing data may be added to or
overwritten. For example with the HAL adaptor an entity with a given
name will always override any previous entity already defined with the
same name. With the Siren adaptor, it will be added to the list of
existing entities (assuming it’s not a complete duplicate of another
sub-entity).