Avoid usage of instance_eval by defining schemas as methods
While investigating a performance issue in JRuby when using oat in a
specific JRuby configuration, @headius pointed out that the way oat
was using `instance_eval` to apply blocks would have an impact on
the library performance, both on MRI and on JRuby.

In a nutshell, this happens because while every object behaves as
it has a singleton class, the object representing that singleton
class is generated lazily, only when really needed. And this happens
with `instance_eval` --- it causes the singleton class to be created
both on MRI and on JRuby.

This can be confirmed on MRI by using ObjectSpace:

```ruby
>> require 'objspace'
>> ObjectSpace.count_objects[:T_CLASS]
=> 988
>> Object.new.instance_eval { }
=> nil
>> ObjectSpace.count_objects[:T_CLASS]
=> 989
>> Object.new.instance_eval { }
=> nil
>> ObjectSpace.count_objects[:T_CLASS]
=> 990
```

The same happens in JRuby, where a JVM heap visualization tool such
as JVisualVM can observe a new `org.jruby.MetaClass` instance being
created on every call to `instance_eval`.

To avoid this overhead and improve oat's performance, this PR replaces
the `instance_eval` we use to evaluate blocks in the context of a
serializer class with methods that are created from the block.

This means that when a `schema` is registered, we now generate a new
private method from it (e.g. `schema_block_N`) and thus when we're
serializing a given object we can just call this method (or methods,
if several schemas were registered) instead of `instance_eval`.
Because this method is defined on the serializer class, it behaves
exactly as the `instance_eval`, but performance is considerably
increased.

Simple benchmark (using `benchmark-ips` gem):

```ruby
require 'benchmark/ips'
require 'oat'
require 'oat/adapters/hal'

class Foo
  attr_reader :foo, :bar, :baz

  def initialize(**values)
    values.each do |(key, value)|
      respond_to?(key) ? instance_variable_set(:"@#{key}", value) :
raise
    end
  end
end

class FooSerializer < Oat::Serializer
  adapter Oat::Adapters::HAL

  schema do
    property :foo, item.foo
    property :bar, item.bar
    property :baz, item.baz
  end
end

puts "Running with #{RUBY_DESCRIPTION}"

FOO = Foo.new(foo: 1, bar: 2, baz: 'baz')

Benchmark.ips do |benchmark|
  benchmark.time = 30
  benchmark.warmup = 30
  benchmark.report('FooSerializer') { FooSerializer.new(FOO).to_hash }
  benchmark.compare!
end
```

And results on my machine:

* Before
  * MRI 2.4.1:
    `FooSerializer    264.516k (± 7.9%) i/s -      7.890M in  30.032221s`
  * JRuby 9.1.9.0 + invokedynamic:
    `FooSerializer    300.490k (±20.6%) i/s -      8.652M in  30.080662s`

* After
  * MRI 2.4.1:
    `FooSerializer    326.571k (± 5.8%) i/s -      9.771M in  30.035173s`
  * JRuby 9.1.9.0 + invokedynamic:
    `FooSerializer    622.808k (± 3.8%) i/s -     18.680M in  30.039202s`

E.g. this speeds up Oat by 1.23x on MRI and 2.07x on JRuby.

See also https://github.com/jruby/jruby/issues/4596