Improve performance by avoiding use of method_missing in Serializer
Following up the previous commit, another thing that @headius noted
was that Oat's usage of `method_missing` to delegate from the
serializer to the adapter was also impacting performance of the
serialization process.

This `method_missing` operation is used in the Serializer to dispatch
the methods used inside a schema block (e.g. `property`, `link`, etc)
to the currently selected adapter, which implements them.

To improve performance, of this operation, we can "learn" which
methods the adapter implementds on the first `method_missing` calls,
and from them on we define a method on the serializer class that
directly dispatches to the adapter, avoiding the `method_missing`
call.

Note that this works even if the adapter is changed, as we expect
the changed adapter to still answer to the interface (and if it
doesn't, we just get a `NoMethodError`).

Using the same benchmark included in the previous commit, here is
the impact of this optimization:

* Before
  * MRI 2.4.1:
    `FooSerializer    326.571k (± 5.8%) i/s -      9.771M in  30.035173s`
  * JRuby 9.1.9.0 + invokedynamic:
    `FooSerializer    622.808k (± 3.8%) i/s -     18.680M in  30.039202s`

* After
  * MRI 2.4.1:
    `FooSerializer    374.500k (± 4.1%) i/s -     11.246M in  30.084775s`
  * JRuby 9.1.9.0 + invokedynamic:
    `FooSerializer      1.058M (± 4.7%) i/s -     31.695M in  30.042851s`

This speeds up Oat by 1.15x on MRI and 1.70x on JRuby. Note that these
numbers are relative to the previous commit; relative to current master
both improvements yield 1.41x on MRI and 3.52x on JRuby.

See also https://github.com/jruby/jruby/issues/4596