Allow for more sophisticated pluralization
While removing the dependency on ActiveSupport is good, it sometimes results in suboptimal results.

Therefore, allow the hosting application to provide the String with a pluralize-method. This could come from ActiveSupport or a custom method, tailored towards the need of the application. 

If this method is not available, continue with the naïve implementation that works most of the time.